#43 2to3: CleanTrim Points - including conversion to wizard.
